+++ src/schemaGenerator.js
@@ -279,9 +279,23 @@ function assertResolveFunctionsPresent(schema, resolverValidationOptions = {}) {
   const {
     requireResolversForArgs = true,
     requireResolversForNonScalar = true,
+    requireResolversForAllFields = false,
   } = resolverValidationOptions;
 
+  if (requireResolversForAllFields && (!requireResolversForArgs || !requireResolversForNonScalar)) {
+    throw new TypeError(
+      'requireResolversForAllFields takes precedence over the more specific assertions. ' +
+      'Please configure either requireResolversForAllFields or requireResolversForArgs / ' +
+      'requireResolversForNonScalar, but not a combination of them.'
+    );
+  }
+
   forEachField(schema, (field, typeName, fieldName) => {
+    // requires a resolve function for *every* field.
+    if (requireResolversForAllFields) {
+      expectResolveFunction(field, typeName, fieldName);
+    }
+
     // requires a resolve function on every field that has arguments
     if (requireResolversForArgs && field.args.length > 0) {
       expectResolveFunction(field, typeName, fieldName);